Add voice-customer blueprint and update brand endpoints
