+++ backend/app.py
@@ -119,9 +119,12 @@ from data_summary.file_utils import detect_extension
 from data_summary.config import get_azure_openai_config
 from data_summary.llm import PandasAIClient
 from data_summary.summarize import create_description
+
 from routes.report_jobs import bp as jobs_bp
 from routes.organizations import bp as organizations
 from routes.upload_source_document import bp as upload_source_document
+from routes.voice_customer import bp as voice_customer
+
 from _secrets import get_secret
 
 # Suppress Azure SDK logs (including Key Vault calls)
@@ -227,6 +230,7 @@ def _load_secrets_once():
 app.register_blueprint(jobs_bp)
 app.register_blueprint(organizations)
 app.register_blueprint(upload_source_document)
+app.register_blueprint(voice_customer)
 
 
 def handle_auth_error(func):
@@ -4972,134 +4976,6 @@ def update_url():
     except Exception as e:
         logger.exception(f"Unexpected error in modify_url: {e}")
         return create_error_response("Internal Server Error", 500)
-
-
-@app.route("/api/voice-customer/brands", methods=["POST"])
-def create_brand():
-    """
-    Handles the creation of a new brand.
-
-    Expects a JSON payload with the following required fields:
-        - brand_name (str): The name of the brand.
-        - brand_description (str): A description of the brand.
-        - organization_id (int or str): The ID of the associated organization.
-
-    Returns:
-        - On success: A JSON response with the created brand data and HTTP status 201.
-        - On failure: A JSON error response with an appropriate error message and HTTP status code.
-    """
-    data = request.get_json()
-    if not data:
-        return create_error_response("No JSON data provided", 400)
-    required_fields = ["brand_name", "organization_id"]
-    missing_fields = [field for field in required_fields if field not in data]
-    if missing_fields:
-        return create_error_response(
-            f"Missing required fields: {', '.join(missing_fields)}", 400
-        )
-    try:
-        brand_name = data["brand_name"]
-        brand_description = data.get("brand_description", "")
-        organization_id = data["organization_id"]
-
-        result = create_new_brand(
-            brand_name=brand_name,
-            brand_description=brand_description,
-            organization_id=organization_id,
-        )
-        return create_success_response(result, 201)
-    except Exception as e:
-        return create_error_response(f"Error creating brand: {str(e)}", 500)
-
-
-@app.route("/api/voice-customer/brands/<brand_id>", methods=["DELETE"])
-def delete_brand(brand_id):
-    """
-    Deletes a brand by its ID.
-
-    Args:
-        brand_id (str or int): The unique identifier of the brand to delete.
-
-    Returns:
-        Response: A success response with the result of the deletion and HTTP status 200,
-                  or an error response with an appropriate message and status code.
-
-    Raises:
-        Exception: If an error occurs during the deletion process.
-    """
-    if not brand_id:
-        return create_error_response("Brand ID is required", 400)
-    try:
-        response = delete_brand_by_id(brand_id)
-        return create_success_response(response, 200)
-    except Exception as e:
-        return create_error_response(f"Error deleting brand: {str(e)}", 500)
-
-
-@app.route(
-    "/api/voice-customer/organizations/<organization_id>/brands", methods=["GET"]
-)
-def get_brands(organization_id):
-    """
-    Retrieve brands associated with a given organization.
-
-    Args:
-        organization_id (str or int): The unique identifier of the organization.
-
-    Returns:
-        Response: A success response containing the list of brands (HTTP 200),
-                  or an error response with an appropriate message and status code (HTTP 400 or 500).
-
-    Raises:
-        Exception: If an unexpected error occurs during brand retrieval.
-    """
-    if not organization_id:
-        return create_error_response("Organization ID is required", 400)
-    try:
-        brands = get_brands_by_organization(organization_id)
-        return create_success_response(brands, 200)
-    except Exception as e:
-        return create_error_response(f"Error retrieving brands: {str(e)}", 500)
-
-
-@app.route("/api/voice-customer/brands/<brand_id>", methods=["PATCH"])
-def update_brand(brand_id):
-    """
-    Updates the details of a brand with the specified brand_id.
-    Expects a JSON payload with the following required fields:
-        - brand_name (str): The new name of the brand.
-        - brand_description (str): The new description of the brand.
-    Args:
-        brand_id (int or str): The unique identifier of the brand to update.
-    Returns:
-        Response: A JSON response indicating success with the updated brand data and HTTP 200 status,
-                  or an error message with the appropriate HTTP status code if the request is invalid
-                  or an error occurs during the update process.
-    """
-    data = request.get_json()
-    if not data:
-        return create_error_response("No JSON data provided", 400)
-
-    required_fields = ["brand_name", "brand_description"]
-    missing_fields = [field for field in required_fields if field not in data]
-    if missing_fields:
-        return create_error_response(
-            f"Missing required fields: {', '.join(missing_fields)}", 400
-        )
-
-    try:
-        brand_name = data["brand_name"]
-        brand_description = data["brand_description"]
-
-        result = update_brand_by_id(
-            brand_id=brand_id,
-            brand_name=brand_name,
-            brand_description=brand_description,
-        )
-        return create_success_response(result, 200)
-    except Exception as e:
-        return create_error_response(f"Error updating brand: {str(e)}", 500)
-
 
 @app.route("/api/voice-customer/products", methods=["POST"])
 def create_product():
